moves addPrayer mutation
